Changed API of `BindingView<B, S>` to better handle view bindings
The `::bind()` api was kind of awry with requiring the user to call a
passed lambda (which made no sense really) that would set the class
binding. This could be replaced by an `onStart()` function that would
require the binding to be set much like `setContentView()` works for
activities.

Alternatively, I could keep the `bind()` method, but without the lambda
and this time requiring it to return the binding itself instead of the
`android.view.View`. The reason this wasn't originally the case is that
the interface `ViewBinding` is not available by default, requiring the
package to be included (which I didn't know). Now it's possible and the
API looks much tidier. I may change it to be more like Activity in this
regard, however. First I'll try to see if we can make instantiating the
binding a little slimmer still.