Adds domain modelling for bills in home
Adds `CoroutinesExecutor` which can hold coroutines executions (duh) and
be registered as a destroyable to clear the call upon `onDestroy()`. It
is fine to call for registration deliberately since internally it's a
set and if we try to push it more than once it'll just be ignored.

And yes, we'll use coroutines here.