Corrects usage of diff util callback to properly calculate item changes
Previously we were considering changes in content to translate into
changing the item in the list completely, which does not hold true for
more complex identifiable data such as these with an `id` and whose
content may change over time.

To properly accomodate that scenario, an `IdentityMatcher<T>` was added
to the `ListDiffCallback` API (with was the previous `ListDiffUtil`) so
that we can define the behavior at will. The default behavior is still
to consider a change in value to be a change in item as well.

Adds tests to assert things are working as expected.