Modifies Controller API and disassembles it's definition from Fragment
Separates the controlling logic to an external engine, but keeps logic
related to linking controller functionality to Fragment API. Creates an
interface that defines a responsibility boundary to what belongs to a
controller in a more general sense. Not only the the engine complies to
that definition, forthcoming possible controller points should also
comply.

This definition is still unstable and should see further modifications
that better harmonize it's general intended use.